+++ example/bin/main.dart
@@ -3,6 +3,7 @@ import 'package:example/injector/injector.dart';
 import 'package:injectable/injectable.dart';
 
 void main(List<String> arguments) {
-  configInjector(environmentFilter: NoEnvOrContainsAny({prod.name, platformMobile.name}));
+  configInjector(environmentFilter: NoEnvOrContains(platformMobile.name));
+
   print(getIt<Service>().environments);
 }