+++ example/lib/injector/Service.dart
@@ -1,4 +1,3 @@
-
 abstract class Service {
   Set<String> get environments;
 }