+++ injectable_generator/lib/dependency_config.dart
@@ -180,8 +180,12 @@ class ImportableType {
     return "$namePrefix$name$typeArgs";
   }
 
-  String getDisplayName(Set<ImportableType> prefixedTypes, {bool includeTypeArgs = true}) {
-    return prefixedTypes?.lookup(this)?.fullName(includeTypeArgs: includeTypeArgs) ?? fullName(includeTypeArgs: includeTypeArgs);
+  String getDisplayName(Set<ImportableType> prefixedTypes,
+      {bool includeTypeArgs = true}) {
+    return prefixedTypes
+            ?.lookup(this)
+            ?.fullName(includeTypeArgs: includeTypeArgs) ??
+        fullName(includeTypeArgs: includeTypeArgs);
   }
 
   String get importName => "'$import' ${prefix != null ? 'as $prefix' : ''}";
@@ -212,17 +216,16 @@ class ImportableType {
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
-          other is ImportableType &&
-              runtimeType == other.runtimeType &&
-              identity == other.identity &&
-              ListEquality().equals(typeArguments, other.typeArguments);
+      other is ImportableType &&
+          runtimeType == other.runtimeType &&
+          identity == other.identity &&
+          ListEquality().equals(typeArguments, other.typeArguments);
 
   @override
   int get hashCode =>
       import.hashCode ^ name.hashCode ^ ListEquality().hash(typeArguments);
 
-  Map<String, dynamic> toJson() =>
-      {
+  Map<String, dynamic> toJson() => {
         "name": name,
         "import": import,
         if (typeArguments?.isNotEmpty == true)